+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix errors appearing when user switch to "User consolidation" pane. (UIEH-1362)
 * Fix errors appearing when user switch to "User consolidation" pane. (UIEH-1362)
 * Settings > Usage Consolidation naively parses start-month values. (UIEH-1371)
+* Settings: Add `PaneCloseLink` to the `Assigned users` pane. (UIEH-1372)
 
 ## [8.0.3] (https://github.com/folio-org/ui-eholdings/tree/v8.0.3) (2023-03-30)